+++ Makefile
@@ -1,5 +1,5 @@
 # Run make new_version after changing this version
-VERSION=1.4.1
+VERSION=1.4.2
 
 .PHONY: compile_app_binary
 compile_app_binary:
