--- app/onepassword_events_api/appserver/static/javascript/setup_page_141.js
@@ -1,17 +1,0 @@
-"use strict";
-
-var app_name = "./onepassword_events_api";
-
-require.config({
-  paths: {
-    myApp: "../app/" + app_name + "/javascript/views/app_141",
-    react: "../app/" + app_name + "/javascript/vendor/react.production.min",
-    ReactDOM:
-      "../app/" + app_name + "/javascript/vendor/react-dom.production.min",
-  },
-  scriptType: "module",
-});
-
-require(["react", "ReactDOM", "myApp"], function (react, ReactDOM, myApp) {
-  ReactDOM.render(myApp, document.getElementById("main_container"));
-});
