+++ app/onepassword_events_api/default/app.conf
@@ -1,6 +1,6 @@
 [install]
 is_configured = 0
-build = 141
+build = 142
 
 [package]
 id = onepassword_events_api
@@ -13,7 +13,7 @@ setup_view = setup_page_dashboard
 [launcher]
 author =  1Password - Insights
 description = Gain insight into your 1Password activity with 1Password Events API for Splunk.
-version = 1.4.1
+version = 1.4.2
 
 [triggers]
 reload.events_reporting = simple
